Struct prusst::EvtoutIrq

pub struct EvtoutIrq {
    // some fields omitted
}

Synchronization primitive that can be used to wait for an event out.

Methods

impl EvtoutIrq

fn wait(&self) -> u32

Waits until the associated event out is triggered.

Panics

This function should not panic as long as the UIO module is loaded, which is theoretically guaranteed at this point since Pruss could not have been created otherwise.

fn get_evtout(&self) -> Evtout

Returns the associated event out.