Struct prusst::Intc

pub struct Intc {
    // some fields omitted
}

The PRU interrupt controller.

Methods

impl Intc

fn map_interrupts(&mut self, interrupts: &IntcConfig)

Maps PRU interrupts according to the provided configuration.

fn send_sysevt(&self, sysevt: Sysevt)

Triggers a system event.

fn clear_sysevt(&self, sysevt: Sysevt)

Clears a system event.

fn enable_sysevt(&self, sysevt: Sysevt)

Enables a system event.

fn disable_sysevt(&self, sysevt: Sysevt)

Disables a system event.

fn enable_host<T: Into<Host>>(&self, host: T)

Enables or re-enables a host interrupt.

Beware: calling this function before the triggering system event was cleared will trigger the host interrupt again.

fn disable_host<T: Into<Host>>(&self, host: T)

Disables a host interrupt.

fn register_irq(&self, e: Evtout) -> EvtoutIrq

Returns a synchronization primitive for event out host interrupts.

Important: this function should be called before any corresponding event out is triggered.

Panics

This function should not panic provided that the uio_pruss kernel module is loaded, which is theoretically guaranteed at this point since Pruss could not have been created otherwise.