Struct prusst::PruLoader

pub struct PruLoader {
    // some fields omitted
}

PRU instruction code loader.

Methods

impl PruLoader

fn load_code<R: Read>(&mut self, code: &mut R) -> Result<PruCode>

Loads a binary of opcodes to the PRU without executing it.

This function proceeds as follows:

a soft PRU reset is forced,

the code is written to the PRU instruction RAM.

The code can be subsequently started and stopped using the returned PruCode handle.

Errors

IO errors that may occur while reading the buffer are forwarded. If the buffer cannot be read entirely because the code does not fit into the instruction RAM, an error of the kind ErrorKind::InvalidInput is returned.