Struct prusst::util::VolatileCell

pub struct VolatileCell<T> {
    // some fields omitted
}

A convenient zero-overhead wrapper for volatile fields that implement Copy and Clone.

This allows writing and reading volatile fields without resorting to unsafe.

An important difference with another popular implementation of VolatileCell is that interior mutability is not allowed. Although it may at first appear logical to allow a volatile Send and Sync type to provide interior mutability arguing that the value may anyway be changed by another process, this may wrongly lead the user into believing that volatile values can be used used for inter-thread communication (see A. D. Robinson's essay: "Volatile: Almost Useless for Multi-Threaded Programming").

Methods

impl<T> VolatileCell<T>

fn new(value: T) -> VolatileCell<T>

Creates a new VolatileCell containing the given value.

fn get(&self) -> T

Returns a copy of the contained value.

fn set(&mut self, value: T)

Sets the contained value.